pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
version= "0.0.0"
name = "ragnarok-stmloader"
authors = [
    {name = "Laurent Bonnet", email = "laurent.woolcap@free.fr"}
]
description = "Usage of STM32 Bootloader command over scaffold v1.1 from ledger"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "intelhex",
    "donjon-scaffold",
    "progressbar2",
    "typer",
    "cerberus",
    "pyyaml"
]

classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Natural Language :: English",
    "Operating System :: OS Independent",
]

[project.scripts]
stmloader = "stm32boot.__main__:cli"

[project.urls]
Home = "https://github.com/durufle/stm32boot"
BugTracker = "https://github.com/durufle/stm32boot/issues"
SourceCode = "https://github.com/durufle/stm32boot"

[tool.setuptools.packages.find]
exclude = ['firmwares*', 'tests*', 'venv*', 'docs*', 'daughter-board*', 'scripts*']

[tool.setuptools]
include-package-data = true

[tool.pylint.master]
ignore="conf.py, ceval_stm32l4xx.py, test_loader_cli_args.py"

[tool.pylint.format]
max-line-length = 120
disable = "W0104, R0902, R0904, W0612"



